Implementera skjut funktioner

+++ core/src/com/mygdx/game/entites/entitiycomponents/DamageComponent.java
@@ -3,17 +3,17 @@ package com.mygdx.game.entites.entitiycomponents;
 import com.badlogic.ashley.core.Component;
 
 public class DamageComponent implements Component{
-	private Double _damage;
+	private float _damage;
 
-	public DamageComponent(Double damage) {
+	public DamageComponent(float damage) {
 		setDamage(damage);
 	}
 
-	public Double getDamage() {
+	public float getDamage() {
 		return _damage;
 	}
 
-	public void setDamage(Double damage) {
+	public void setDamage(float damage) {
 		_damage = damage;
 	}
